+++ CHANGES.md
@@ -7,6 +7,7 @@
 * Built using Avro 1.9.2 (#104)
 * Add support for Java 13
 * Add support for testing multiple Kotlin versions
+* Update plugin's own build to address some deprecation warnings of APIs being removed in Gradle 7
 
 ## 0.18.0
 * Use reproducible file order for plugin archives