+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## Unreleased
-* Add support for Gradle 6.0-6.2 (#101)
+* Add support for Gradle 6.0-6.2.2 (#101)
 * Drop support for Gradle versions prior to 5.1
 * Update version of kotlin plugin in tests/example
 * Built using Avro 1.9.2 (#104)